Move handlers to an internal package (#95)
